requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14
drf-spectacular>=0.27  # For OpenAPI docs
psycopg2-binary>=2.9   # Postgres adapter
django-environ>=0.11   # 12-Factor App config
gunicorn>=21.2         # Production WSGI server
whitenoise>=6.6        # Static file serving
django-cors-headers>=4.3  # CORS handling for frontend